Implement Phoenix 1.3 Namespacing Recognition
* Phoenix 1.3 namespaces now supported with phx.swagger.generate.
* Version bump to 0.6.5.
* Fixes Issue 117.

Between Phoenix 1.3 RCs and the Phoenix 1.3 release, a
modification was made to the Web namespace: instead of
the namespace being {ProjectName}.Web, it was changed to
{ProjectName}Web. The namespace for the router and endpoint
now correctly defaults to the correct 1.3 namespace.

Phoenix 1.2 applications are still supported by providing the
router and endpoint to phx.swagger.generate using the -r and -e
switches.